pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "werkzeug>=3.1.3",
    "sqlalchemy>=2.0.41",
    "stripe>=12.2.0",
    "pillow>=11.2.1",
    "bcrypt>=4.3.0",
    "python-dotenv>=1.1.1",
    "flask-socketio>=5.5.1",
    "eventlet>=0.40.1",
    "libtorrent>=2.0.11",
    "requests>=2.32.4",
    "aiohttp>=3.12.15",
    "anilistpython>=0.1.3",
    "trafilatura>=2.0.0",
    "beautifulsoup4>=4.13.4",
    "urllib3>=2.5.0",
    "flask-dance>=7.1.0",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
]
